delete dataset fix and collaborator modal update fix
