+++ ui/src/js/components/shared/header/actionsSection/collaborators/modal/CollaboratorsModal.scss
