+++ ui/src/js/mutations/DeleteDatasetMutation.js
@@ -31,7 +31,6 @@ function sharedUpdater(store, parentId, deletedId, connectionKey) {
   }
 }
 
-
 export default function DeleteRemoteDatasetMutation(
   datasetName,
   owner,
